Added the photospicker and transition into the view
User can now click an Add button to go straight to the Add/Edit view. When they do, they are immediately presented with a photos picker to choose the image they want to add.

Logic also added to deal with cancelling out of the photos picker, and hiding the main view details unless the user chooses something.

Signed-off-by: Pete Wright <peterjwright+uk@gmail.com>