Fixing the layout of the list and thumbnails
Added an in memory container for the data to handle previews. Also set up the preview on the list view so that we create some stub data.

Modified the list grid layout so that it’s 2 columns instead of 3 - it was too hard to see what’s going on otherwise.

Signed-off-by: Pete Wright <peterjwright+uk@gmail.com>