Got the sort buttons to change highlights
We now track the sort order on each of the home page and list pages, and update the formatting of the sort buttons accordingly to show the user what is active.

Does beg the question though as to whether or not this needs to be global across both views, or independent.

Signed-off-by: Pete Wright <peterjwright+uk@gmail.com>